These changes allow to install pyiga with Python 3.12
1) running setup.py itself is deprecated; since Python 3.12,
setuptools is not bundled anymore, so setup.py can be executed only
after installation of setuptools. Instead, it is recommended to
install using pip, which automatically installs setuptools in
isolated context (PEP 518). This requires the minimal pyproject.toml
file. Using setup.py to configure setuptools is not deprecated.

2) more recent versions of numpy do not provide np.int_t anymore.

3) cython complains that implicit noexcept was deprecated, in
some instances that actually broke the code with "Signature not
compatible with previous declaration" error messages.